TaxCategoryDraft to interface. #942
