+++ sphere-models/src/main/java/io/sphere/sdk/taxcategories/TaxCategoryDraftImpl.java
@@ -1,0 +1,39 @@
+package io.sphere.sdk.taxcategories;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.sphere.sdk.models.Base;
+
+import javax.annotation.Nullable;
+import java.util.List;
+
+final class TaxCategoryDraftImpl extends Base implements TaxCategoryDraft {
+    private final String name;
+    @Nullable
+    private final String description;
+    private final List<TaxRate> taxRates;
+
+    @JsonCreator
+    TaxCategoryDraftImpl(final String name, @Nullable final String description, final List<TaxRate> taxRates) {
+        this.name = name;
+        this.description = description;
+        this.taxRates = taxRates;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    @Nullable
+    public String getDescription() {
+        return description;
+    }
+
+    @Override
+    @JsonProperty("rates")
+    public List<TaxRate> getTaxRates() {
+        return taxRates;
+    }
+}
